# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ACUM Search API
  version: 0.1.0
  description: |
    Open source REST wrapper around the ACUM (אקו"ם) public music rights search database.

    **Not affiliated with or endorsed by ACUM.** Wraps only unauthenticated public endpoints.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    url: https://github.com/yanirclsr/acum-api

servers:
  - url: http://localhost:3000
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      tags: [meta]
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/search:
    get:
      summary: Search works and artists
      operationId: searchWorks
      tags: [search]
      parameters:
        - name: q
          in: query
          required: true
          description: Primary search text (Hebrew or transliterated)
          schema:
            type: string
            example: לילה
        - name: by
          in: query
          schema:
            type: string
            enum: [title, artist, composer, performer, album, catalog, number]
            default: title
        - name: artist
          in: query
          description: Optional secondary filter — narrow results by performer/artist name
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
            enum: [partial, exact]
            default: partial
        - name: sort
          in: query
          schema:
            type: string
            enum: [alphabetical, reverse]
            default: alphabetical
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 30
            default: 10
        - name: type
          in: query
          description: Return works (default) or artists
          schema:
            type: string
            enum: [works, artists]
            default: works
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/UpstreamError"

  /api/works/{id}:
    get:
      summary: Get work detail with versions
      operationId: getWork
      tags: [works]
      parameters:
        - name: id
          in: path
          required: true
          description: Full work ID (e.g. 1579291)
          schema:
            type: string
            pattern: '^\d{7,10}$'
      responses:
        "200":
          description: Work detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamError"

  /api/works/{id}/versions/{versionId}:
    get:
      summary: Get version detail (includes ISWC, ISRC, full creator list)
      operationId: getVersion
      tags: [works]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          description: Full version ID (e.g. 1579291001)
          schema:
            type: string
      responses:
        "200":
          description: Version detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/artists/search:
    get:
      summary: Search artists/composers by name
      operationId: searchArtists
      tags: [artists]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
            enum: [partial, exact]
            default: partial
        - name: sort
          in: query
          schema:
            type: string
            enum: [alphabetical, reverse]
            default: alphabetical
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 30
            default: 10
      responses:
        "200":
          description: Artist results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistSearchResult"

  /api/artists/{id}/works:
    get:
      summary: Get all works by an artist (by IpBaseNumber)
      operationId: getArtistWorks
      tags: [artists]
      parameters:
        - name: id
          in: path
          required: true
          description: Artist IpBaseNumber (e.g. I-000151826-7)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          schema:
            type: string
            enum: [alphabetical, reverse]
            default: alphabetical
      responses:
        "200":
          description: Works by this artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"

components:
  schemas:
    Creator:
      type: object
      properties:
        ipBaseNumber:
          type: string
          example: I-000151826-7
        nameHebrew:
          type: string
          example: שלמה ארצי
        nameEnglish:
          type: string
          example: SHLOMO ARTZI
        roleCode:
          type: string
          enum: [A, C, CA, AR, AT, E, SE, CO, PA]
          description: A=Author, C=Composer, CA=Both, AR=Arranger, AT=Translator, E=Publisher
        protectionStatus:
          type: string
          enum: ["1", "2", "3"]
          description: "1=ACUM protected, 2=Not specified, 3=Public domain"

    Performer:
      type: object
      properties:
        id:
          type: string
        nameHebrew:
          type: string
        nameEnglish:
          type: string

    Work:
      type: object
      properties:
        id:
          type: string
          example: "1579291"
        versionId:
          type: string
          example: "1579291001"
        workNumber:
          type: string
        versionNumber:
          type: string
        titleHebrew:
          type: string
          example: לילה
        titleEnglish:
          type: string
          example: LAYLA
        isForeign:
          type: boolean
        pool:
          type: string
          enum: [local, foreign]
        category:
          type: string
          enum: [musical, literature, medley, translated]
        registrationDate:
          type: string
        publicationDate:
          type: string
        composers:
          type: array
          items:
            $ref: "#/components/schemas/Creator"
        authors:
          type: array
          items:
            $ref: "#/components/schemas/Creator"
        arrangers:
          type: array
          items:
            $ref: "#/components/schemas/Creator"
        publishers:
          type: array
          items:
            $ref: "#/components/schemas/Creator"
        performer:
          $ref: "#/components/schemas/Performer"
        iswc:
          type: string
          example: T3385316923
        isrc:
          type: string
        duration:
          type: string
          example: "03:13"
        versionCount:
          type: integer
        acumUrl:
          type: string

    WorkDetail:
      allOf:
        - $ref: "#/components/schemas/Work"
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/Work"

    Artist:
      type: object
      properties:
        ipBaseNumber:
          type: string
          example: I-000151826-7
        nameHebrew:
          type: string
        nameEnglish:
          type: string
        workCount:
          type: integer
        versionCount:
          type: integer
        profession:
          type: string
          enum: [composer, author, composer_author, publisher, other]
        joinYear:
          type: string
        ipnNumber:
          type: string
        caeNumber:
          type: string

    SearchResult:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Work"

    ArtistSearchResult:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Artist"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded (60 req/min/IP)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UpstreamError:
      description: ACUM service error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"